refactor: move out handlers and commands to separate file
